requirements.txt.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
python-dotenv
pydantic
jinja2
watchfiles
